refactor: extract homepage project card into the ProjectCard component
